--- a/broker/src/main/java/io/zeebe/broker/Broker.java
+++ b/broker/src/main/java/io/zeebe/broker/Broker.java
@@ -45,6 +45,7 @@ import io.zeebe.engine.processor.workflow.EngineProcessors;
 import io.zeebe.engine.processor.workflow.message.command.SubscriptionCommandSender;
 import io.zeebe.engine.state.ZeebeState;
 import io.zeebe.logstreams.log.LogStream;
+import io.zeebe.logstreams.storage.atomix.ZeebeIndexAdapter;
 import io.zeebe.protocol.impl.encoding.BrokerInfo;
 import io.zeebe.transport.ServerTransport;
 import io.zeebe.transport.TransportFactory;
@@ -56,7 +57,9 @@ import io.zeebe.util.sched.ActorControl;
 import io.zeebe.util.sched.ActorScheduler;
 import io.zeebe.util.sched.clock.ActorClock;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
@@ -87,6 +90,7 @@ public final class Broker implements AutoCloseable {
   private EmbeddedGatewayService embeddedGatewayService;
   private ServerTransport serverTransport;
   private BrokerHealthCheckService healthCheckService;
+  private Map<Integer, ZeebeIndexAdapter> partitionIndexes;
 
   public Broker(final SystemContext systemContext) {
     this.brokerContext = systemContext;
@@ -189,6 +193,22 @@ public final class Broker implements AutoCloseable {
 
   private AutoCloseable atomixCreateStep(final BrokerCfg brokerCfg) {
     atomix = AtomixFactory.fromConfiguration(brokerCfg);
+
+    final var partitionGroup =
+        (RaftPartitionGroup)
+            atomix.getPartitionService().getPartitionGroup(AtomixFactory.GROUP_NAME);
+
+    partitionIndexes = new HashMap<>();
+    final var logIndexDensity = brokerCfg.getData().getLogIndexDensity();
+    partitionGroup.getPartitions().stream()
+        .map(RaftPartition.class::cast)
+        .forEach(
+            raftPartition -> {
+              final var zeebeIndex = ZeebeIndexAdapter.ofDensity(logIndexDensity);
+              partitionIndexes.put(raftPartition.id().id(), zeebeIndex);
+              raftPartition.setJournalIndexFactory(() -> zeebeIndex);
+            });
+
     return () ->
         atomix.stop().get(brokerContext.getStepTimeout().toMillis(), TimeUnit.MILLISECONDS);
   }
@@ -296,8 +316,9 @@ public final class Broker implements AutoCloseable {
     final StartProcess partitionStartProcess = new StartProcess("Broker-" + nodeId + " partitions");
 
     for (final RaftPartition owningPartition : owningPartitions) {
+      final var partitionId = owningPartition.id().id();
       partitionStartProcess.addStep(
-          "partition " + owningPartition.id().id(),
+          "partition " + partitionId,
           () -> {
             final ZeebePartition zeebePartition =
                 new ZeebePartition(
@@ -308,6 +329,7 @@ public final class Broker implements AutoCloseable {
                     scheduler,
                     brokerCfg,
                     commandHandler,
+                    partitionIndexes.get(partitionId),
                     createFactory(topologyManager, clusterCfg, atomix, managementRequestHandler));
             scheduleActor(zeebePartition);
             healthCheckService.registerMonitoredPartition(