--- a/broker/src/main/java/io/zeebe/broker/system/configuration/DataCfg.java
+++ b/broker/src/main/java/io/zeebe/broker/system/configuration/DataCfg.java
@@ -25,6 +25,7 @@ public final class DataCfg implements ConfigurationEntry {
   private Duration snapshotPeriod = Duration.ofMinutes(15);
 
   private int maxSnapshots = 3;
+  private int logIndexDensity = 100;
 
   @Override
   public void init(final BrokerCfg globalConfig, final String brokerBase) {
@@ -68,6 +69,14 @@ public final class DataCfg implements ConfigurationEntry {
     this.maxSnapshots = maxSnapshots;
   }
 
+  public int getLogIndexDensity() {
+    return logIndexDensity;
+  }
+
+  public void setLogIndexDensity(int logIndexDensity) {
+    this.logIndexDensity = logIndexDensity;
+  }
+
   @Override
   public String toString() {
     return "DataCfg{"
@@ -81,6 +90,8 @@ public final class DataCfg implements ConfigurationEntry {
         + '\''
         + ", maxSnapshots="
         + maxSnapshots
+        + ", logIndexDensity="
+        + logIndexDensity
         + '}';
   }
 }