--- a/logstreams/src/main/java/io/zeebe/logstreams/impl/log/LogStorageAppender.java
+++ b/logstreams/src/main/java/io/zeebe/logstreams/impl/log/LogStorageAppender.java
@@ -56,7 +56,6 @@ public final class LogStorageAppender extends Actor {
     this.logStorage = logStorage;
     this.writeBufferSubscription = writeBufferSubscription;
     this.maxAppendBlockSize = maxBlockSize;
-
     appendBackpressureMetrics = new AppendBackpressureMetrics(partitionId);
 
     final boolean isBackpressureEnabled =