--- a/logstreams/src/main/java/io/zeebe/logstreams/storage/atomix/AtomixLogStorage.java
+++ b/logstreams/src/main/java/io/zeebe/logstreams/storage/atomix/AtomixLogStorage.java
@@ -19,24 +19,28 @@ public class AtomixLogStorage implements LogStorage {
   private final AtomixAppenderSupplier appenderSupplier;
 
   private boolean opened;
+  private final ZeebeIndexMapping zeebeIndexMapping;
 
   public AtomixLogStorage(
+      final ZeebeIndexMapping zeebeIndexMapping,
       final AtomixReaderFactory readerFactory,
       final AtomixLogCompactor logCompacter,
       final AtomixAppenderSupplier appenderSupplier) {
+    this.zeebeIndexMapping = zeebeIndexMapping;
     this.readerFactory = readerFactory;
     this.logCompacter = logCompacter;
     this.appenderSupplier = appenderSupplier;
   }
 
-  public static AtomixLogStorage ofPartition(final RaftPartition partition) {
+  public static AtomixLogStorage ofPartition(
+      final ZeebeIndexMapping zeebeIndexMapping, final RaftPartition partition) {
     final var server = new AtomixRaftServer(partition.getServer());
-    return new AtomixLogStorage(server, server, server);
+    return new AtomixLogStorage(zeebeIndexMapping, server, server, server);
   }
 
   @Override
   public LogStorageReader newReader() {
-    return new AtomixLogStorageReader(readerFactory.create());
+    return new AtomixLogStorageReader(zeebeIndexMapping, readerFactory.create());
   }
 
   @Override