--- a/logstreams/src/main/java/io/zeebe/logstreams/storage/atomix/ZeebeIndexMapping.java
+++ b/logstreams/src/main/java/io/zeebe/logstreams/storage/atomix/ZeebeIndexMapping.java
@@ -1,0 +1,20 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Zeebe Community License 1.0. You may not use this file
+ * except in compliance with the Zeebe Community License 1.0.
+ */
+package io.zeebe.logstreams.storage.atomix;
+
+public interface ZeebeIndexMapping {
+
+  /**
+   * Takes a position and returns the corresponding index, where this entry with the given position
+   * is stored.
+   *
+   * @param position
+   * @return
+   */
+  long lookupPosition(long position);
+}