--- a/logstreams/src/test/java/io/zeebe/logstreams/impl/log/LogStorageAppenderTest.java
+++ b/logstreams/src/test/java/io/zeebe/logstreams/impl/log/LogStorageAppenderTest.java
@@ -68,6 +68,7 @@ public final class LogStorageAppenderTest {
             .initialPartitionId(0)
             .build();
     final var subscription = dispatcher.openSubscription("log");
+
     appender =
         new LogStorageAppender(
             "appender", PARTITION_ID, logStorage, subscription, MAX_FRAGMENT_SIZE);