… logstreams/src/test/java/io/zeebe/logstreams/storage/atomix/ZeebeIndexTest.java
… logstreams/src/test/java/io/zeebe/logstreams/util/AtomixLogStorageRule.java