## Description

**NOTE:** The build will fail until https://github.com/zeebe-io/atomix/pull/165 is merged

 In order to improve seek performance and reduce log deletion time,
 we introducing a new ZeebeIndex. This index wraps the Atomix
 SparseJournalIndex and builds an own map of position to index.

 With this index it is easily and in a performant way possible to look
 up an index/address for a given position. This look up is used
 for example on deletion instead of the old seek strategy. This treats
 time against space, which means we need to use more space in memory but
 are fast to find an index/position.

 The ZeebeIndex is on startup injected into the RaftPartition such that
 is filled on leader and follower on appending new entries.

 The underlying data structure needs to be thread safe, since we access
 it on different threads.

Be aware that it is currently just created on Broker startup and injected into the partition before opening them. I plan to refactor the Broker bootstrap and ZeebePartition with https://github.com/zeebe-io/zeebe/issues/3519 and https://github.com/zeebe-io/zeebe/issues/3518 where I then also rewrite this a bit.




### Benchmark

I run a benchmarks with the final changes and it looks like this:


#### General

![run](https://user-images.githubusercontent.com/2758593/75366545-320e0980-58bf-11ea-8c63-bff5ceec815d.png)

We can see that it looks more stable then the latest benchmarks.

#### Resources

![mem](https://user-images.githubusercontent.com/2758593/75366605-43efac80-58bf-11ea-97d2-1fcbb3a10a63.png)

The memory consumption is as expected and shows that the map is correctly cleaned up on every deletion/compaction.

![cpu](https://user-images.githubusercontent.com/2758593/75366652-5964d680-58bf-11ea-9fcd-bcb10d650acd.png)

No changes in CPU usage.

![disk](https://user-images.githubusercontent.com/2758593/75366668-61247b00-58bf-11ea-93dd-191372d2b0ee.png)

The disk space is correctly cleaned up.

![io](https://user-images.githubusercontent.com/2758593/75366681-697cb600-58bf-11ea-8946-d8e27f6a611a.png)

We can see much less IO load or at least much less read bytes per second every 15 min (which is the snapshot interval). Before we had between 60 - 100 Mb/s now we have between 1 to 8 Mb/s.

In benchmarks before I logged how much time it takes to seeking/find the right index. In the first benchmarks it took from 4  to 22 seconds. With introducing the map we can find an approximate index in ~ 1 ms. I also removed the seek to position before deletion since it is not really necessary.


### Benchmark without the map

![run](https://user-images.githubusercontent.com/2758593/75367218-443c7780-58c0-11ea-8dc3-426d3598991d.png)

In this benchmark we seeing much more the recurring throughput pattern. I'm not 100% sure why in this benchmark the throughput is in general higher as in the other. I probably need to investigate that.

![mem](https://user-images.githubusercontent.com/2758593/75367392-8c5b9a00-58c0-11ea-9cf4-80db7c4de71c.png)

Memory consumption looks different without the mapping.

![io](https://user-images.githubusercontent.com/2758593/75367461-a39a8780-58c0-11ea-9977-8b4d254b036f.png)

Here we can see the high spikes of read bytes per s.

## Related issues

<!-- Which issues are closed by this PR or are related -->

closes https://github.com/zeebe-io/zeebe/issues/3918
closes https://github.com/zeebe-io/zeebe/issues/3623


## Pull Request Checklist

- [X] All commit messages match our [commit message guidelines](https://github.com/zeebe-io/zeebe/blob/develop/CONTRIBUTING.md#commit-message-guidelines)
- [X] The submitting code follows our [code style](https://github.com/zeebe-io/zeebe/wiki/Code-Style)
- [X] If submitting code, please run `mvn clean install -DskipTests` locally before committing