@npepinpe 2020-02-26
Haven't had a chance to look but the results look very promising! The only downside I can see is that not compacting now may lead to out of memory issues faster than before, correct? But we can worry about this in a further iteration. Great job! 👍